refactor to clean up main
